+++ src/lib/services/types/index.ts
@@ -9,3 +9,4 @@ export * from "./validator";
 export * from "./wasm";
 export * from "./tx";
 export * from "./move";
+export * from "./stats";